updated nixpkgs properly this time
Context: I had been using a really old version of nixpkgs (tag 23.11)
for a while because I didn't realize that I was targeting the tag
instead of the release branch (nixos-23.11). I probably don't need to
use unstable Chrome/Firefox as well, so I am reverting to the stable
releases: but this time the *updated* stable release. Reversion to the
unstable releases of the browser may or may not happen later depending
on if they get outdated, but most likely not because Chrome was having
an issue where gpu acceleration was turned off.